+++ automated-orders/packages/blockchain-library/src/clients/evm.ts
@@ -19,16 +19,33 @@ import {
   Hex,
   Log,
   PublicClient,
+  TransactionSerializable,
   createPublicClient,
-  createWalletClient,
   decodeEventLog,
   encodeFunctionData,
+  parseTransaction,
+  parseSignature,
+  serializeTransaction,
+  TransactionReceiptNotFoundError,
 } from 'viem';
-import { privateKeyToAccount, signTransaction } from 'viem/accounts';
-
+import { keccak256, hexToBytes } from 'viem';
+import { AwsKmsSigner } from '@warden/aws-kms-signer';
 import { IEvmConfiguration } from '../types/evm/configuration.js';
 import { GasFeeData } from '../types/evm/gas.js';
 import { IEventPollingConfiguration } from '../types/evm/pollingConfiguration.js';
+
+import asn1 from 'asn1.js';
+const { define } = asn1;
+import secp256k1 from 'secp256k1';
+const { ecdsaRecover } = secp256k1;
+
+/* eslint-disable @typescript-eslint/no-explicit-any */
+const EcdsaSigAsnParse = define('EcdsaSig', function(this: any) {
+  this.seq().obj(
+      this.key('r').int(),
+      this.key('s').int(),
+  );
+});
 
 export class EvmClient {
   signer: Hex;
@@ -36,11 +53,9 @@ export class EvmClient {
   events: LRUCache<string, Log>;
   config: Config;
   client: PublicClient;
+  awsKmsSigner: AwsKmsSigner;
 
-  constructor(
-    private configuration: IEvmConfiguration,
-    chain: Chain,
-  ) {
+  constructor(private configuration: IEvmConfiguration, chain: Chain) {
     this.config = createConfig({
       chains: [chain],
       transports: {
@@ -50,7 +65,9 @@ export class EvmClient {
 
     this.client = createPublicClient({
       chain: chain,
-      transport: http(this.configuration.rpcURL),
+      transport: http(this.configuration.rpcURL, {
+        timeout: this.configuration.publicClientTimeout,
+      }),
     });
 
     this.eventsFromBlocks = new Map<string, bigint>();
@@ -61,13 +78,43 @@ export class EvmClient {
       });
     }
 
-    if (this.configuration.callerPrivateKey) {
-      this.signer = privateKeyToAccount(this.configuration.callerPrivateKey as Hex).address;
+    if (this.configuration.awsKmsSignerConfig) {
+      this.awsKmsSigner = new AwsKmsSigner(this.configuration.awsKmsSignerConfig);
     }
   }
 
-  public async broadcastTx(): Promise<void> {
-    // TODO: implementation
+  public async init() {
+    this.signer = (await this.awsKmsSigner.getAddress()) as Hex;
+  }
+
+  public async broadcastTx(rawTransaction: Hex, transactionSignature: Hex) : Promise<void> {
+    const ethRequest = parseTransaction(rawTransaction);
+    const signature = parseSignature(transactionSignature);
+
+    const serialized = serializeTransaction(
+      ethRequest as TransactionSerializable,
+      signature,
+    );
+
+    await this.client.sendRawTransaction({
+      serializedTransaction: serialized,
+    });
+  }
+
+  public async transactionExists(transactionHash: Hex): Promise<boolean> {
+    try {
+      await this.client.getTransactionReceipt({
+        hash: transactionHash,
+      });
+    } catch (error) {
+      if (error instanceof TransactionReceiptNotFoundError) { 
+        return false;
+      }
+
+      throw error;
+    }
+
+    return true;
   }
 
   public async *pollEvents<T>(
@@ -124,7 +171,11 @@ export class EvmClient {
     return (code?.length ?? 0) > 0;
   }
 
-  public async callView<T>(contractAddress: string, functionAbi: AbiFunction, args: unknown[]): Promise<T> {
+  public async callView<T>(
+    contractAddress: string,
+    functionAbi: AbiFunction,
+    args: unknown[],
+  ): Promise<T> {
     const result = await readContract(this.config, {
       abi: [functionAbi],
       functionName: functionAbi.name,
@@ -148,7 +199,12 @@ export class EvmClient {
     return transactionsCount;
   }
 
-  public async getGasFees(from: string, to: string, data: string, value: bigint): Promise<GasFeeData> {
+  public async getGasFees(
+    from: string,
+    to: string,
+    data: string,
+    value: bigint,
+  ): Promise<GasFeeData> {
     const feeData = await estimateFeesPerGas(this.config);
     const gasPrice = await getGasPrice(this.config);
     const gasLimit = await estimateGas(this.config, {
@@ -166,7 +222,11 @@ export class EvmClient {
     };
   }
 
-  public async sendTransaction(contractAddress: string, functionAbi: AbiFunction, args: unknown[]): Promise<void> {
+  public async sendTransaction(
+    contractAddress: string,
+    functionAbi: AbiFunction,
+    args: unknown[],
+  ): Promise<void> {
     const data = encodeFunctionData({
       abi: [functionAbi],
       functionName: functionAbi.name,
@@ -176,26 +236,69 @@ export class EvmClient {
     const gas = await this.getGasFees(this.signer, contractAddress, data, 0n);
     const nonce = await this.getNextNonce(this.signer);
 
-    const wallet = createWalletClient({
-      transport: http(this.configuration.rpcURL),
-      key: this.configuration.callerPrivateKey,
-    });
+    const transaction: TransactionSerializable = {
+      chainId: this.config.chains[0].id,
+      to: contractAddress as Hex,
+      data: data as Hex,
+      gas: gas.gasLimit,
+      maxFeePerGas: gas.maxFeePerGas,
+      maxPriorityFeePerGas: gas.maxPriorityFeePerGas,
+      nonce: nonce,
+      type: 'eip1559',
+      value: 0n,
+    };
 
-    const signedTx = await signTransaction({
-      privateKey: this.configuration.callerPrivateKey! as Hex,
-      transaction: {
-        chainId: this.config.chains[0].id,
-        to: contractAddress as Hex,
-        data: data as Hex,
-        gas: gas.gasLimit,
-        maxFeePerGas: gas.maxFeePerGas,
-        maxPriorityFeePerGas: gas.maxPriorityFeePerGas,
-        nonce: nonce,
-      },
-    });
+    const unsignedSerializedTransaction = serializeTransaction(transaction);
 
-    await wallet.sendRawTransaction({
-      serializedTransaction: signedTx,
+    const transactionHash = keccak256(unsignedSerializedTransaction as Hex);
+    const hashBytes = hexToBytes(transactionHash);
+
+    const signatureDER = await this.awsKmsSigner.signTransactionHash(hashBytes);
+
+    const decodedSignature = EcdsaSigAsnParse.decode(Buffer.from(signatureDER), 'der');
+
+    const r = decodedSignature.r;
+    const s = decodedSignature.s;
+
+    const rBuffer = r.toArrayLike(Buffer, 'be', 32);
+    const sBuffer = s.toArrayLike(Buffer, 'be', 32);
+
+    const signatureCompact = Uint8Array.from(Buffer.concat([rBuffer, sBuffer]));
+
+    let vValue: bigint | undefined;
+
+    for (let recovery = 0; recovery <= 3; recovery++) {
+      const recoveredPublicKey = ecdsaRecover(
+        signatureCompact,
+        recovery,
+        hashBytes,
+        false,
+      );
+      const recoveredAddress =
+        '0x' + keccak256(recoveredPublicKey.slice(1)).slice(-40);
+      if (recoveredAddress.toLowerCase() === this.signer.toLowerCase()) {
+        vValue = BigInt(recovery);
+        break;
+      }
+    }
+
+    if (vValue === undefined) {
+      throw new Error('Failed to recover public key from signature.');
+    }
+
+    const signature = {
+      v: vValue,
+      r: ('0x' + rBuffer.toString('hex')) as Hex,
+      s: ('0x' + sBuffer.toString('hex')) as Hex,
+    };
+
+    const serializedSignedTransaction = serializeTransaction(transaction, signature);
+
+    const serializedSignedTransactionHex = serializedSignedTransaction as Hex;
+
+    await this.client.request({
+      method: 'eth_sendRawTransaction',
+      params: [serializedSignedTransactionHex],
     });
   }