+++ warden/x/async/types/v1beta1/keys.go
@@ -1,0 +1,10 @@
+package v1beta1
+
+const (
+	// ModuleName defines the module name
+	ModuleName = "async"
+)
+
+var (
+	ParamsKey = []byte("p_async")
+)