+++ CHANGELOG.md
@@ -44,12 +44,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * (precompiles) Add an ability for contracts to approve actions
 * (go-client) Return transaction hash from SendWaitTx
 * (keychain-sdk) Log transaction hashes of broadcasted transactions
+* (relayer) #992 Added a service to relay transactions into Ethereum
 
 ### Consensus Breaking Changes
 * (precompiles) Add slinky precompiled contract
 * (x/warden) Add broadcastType to SignRequest
 * (precompiles) Add broadcastType to SignRequest
 * (precompiles) #1077 Change status fields from int32 to enums/uint8 in precompile ABIs for improved type safety and consistency
+* (wardend) Bump IAVL to v1.2.2. Fixes some potential apphash mismatches that happen in some rare cases.
+* (x/async) Scaffold new module with create/read operations
 
 ### Bug Fixes